@@ -11,17 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.0.6",
+    "@angular/animations": "~10.1.6",
     "@angular/cdk": "^9.2.4",
-    "@angular/common": "~9.0.6",
-    "@angular/compiler": "~9.0.6",
-    "@angular/core": "^9.0.7",
-    "@angular/forms": "~9.0.6",
-    "@angular/localize": "^9.0.7",
+    "@angular/common": "~10.1.6",
+    "@angular/compiler": "~10.1.6",
+    "@angular/core": "^10.1.6",
+    "@angular/forms": "~10.1.6",
+    "@angular/localize": "^10.1.6",
     "@angular/material": "^9.2.4",
-    "@angular/platform-browser": "~9.0.6",
-    "@angular/platform-browser-dynamic": "~9.0.6",
-    "@angular/router": "~9.0.6",
+    "@angular/platform-browser": "~10.1.6",
+    "@angular/platform-browser-dynamic": "~10.1.6",
+    "@angular/router": "~10.1.6",
     "@ng-bootstrap/ng-bootstrap": "^6.2.0",
     "@popperjs/core": "^2.4.4",
     "@types/jquery": "^3.3.38",
@@ -41,8 +41,8 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.1001.7",
     "@angular/cli": "^10.1.7",
-    "@angular/compiler-cli": "^9.0.7",
-    "@angular/language-service": "~9.0.6",
+    "@angular/compiler-cli": "^10.1.6",
+    "@angular/language-service": "~10.1.6",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
@@ -57,6 +57,6 @@
     "protractor": "~5.4.3",
     "ts-node": "~8.3.0",
     "tslint": "~5.18.0",
-    "typescript": "~3.7.5"
+    "typescript": "~4.0.3"
   }
 }
