@@ -68,7 +68,7 @@ import { UserListComponent } from './components/user-list/user-list.component';
 })
 export class SharedModule {
   
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<SharedModule> {
     return {
       ngModule: SharedModule,
       providers: [
